Make ignore-scripts configurable
The previous code always invoked npm install with ignore-scripts = true. This
is inconvienent in some cases when it is desired to compile C/C++ modules when
creating the package.

This change provides an ignoreScripts config option, which defaults to true
if not provided, in order to maintain the existing functionality.

An example Gruntfile:
        packageModules: {
            dist: {
                ignoreScripts: false,
                src: 'package.json',
                dest: 'dist'
            }
        },